Fix CI failures

+++ stochtree/bart.py
@@ -1285,7 +1285,7 @@ class BARTModel:
             # the original, which remains alive in this Python scope.
             X_train_cpp = np.asfortranarray(X_train_processed)
             y_train_remapped = y_train - np.min(y_train) if link_is_cloglog else y_train
-            y_train_cpp = np.asfortranarray(y_train_remapped)
+            y_train_cpp = np.asfortranarray(y_train_remapped, dtype=np.float64)
             X_test_cpp = np.asfortranarray(X_test_processed) if self.has_test else None
             basis_train_cpp = (
                 np.asfortranarray(leaf_basis_train) if self.has_basis else None
